// src/app/interfaces/company.ts
export interface CompanyI {
  message: MessageI;
}

export interface MessageI {
  _id: string;
  title: string;
  descripcion: string;
  administrador: string;
  imagen: string;
  telefono1: string;
  telefono2: string;
  telefono3: string;
  logo: string;
  direccion: string;
  email: string;
  facebook: string;
  instagram: string;
  whatsapp: string;
  twitter: string;
  linkedin: string;
  youtube: string;
  solNombre: boolean;
  solDocumento: boolean;
  solCodigo: boolean;
  solDireccion: boolean;
  solTelefono: boolean;
  solTelefono2: boolean;
  solEmail: boolean;
  solIDFamiliar: boolean;
  solImagen: boolean;
  editNombre: boolean;
  editDocumento: boolean;
  editCodigo: boolean;
  editDireccion: boolean;
  editTelefono: boolean;
  editTelefono2: boolean;
  editEmail: boolean;
  editIDFamiliar: boolean;
  editImagen: boolean;
  presentacion: boolean;
  encabezado: boolean;
  clima: boolean;
  color: string;
  aleatorio: boolean;
  cancelar: boolean;
  diaRenovar: string;
  horaRenovar: Date;
  intervaloTurnoAleatorio: string;
  v: number;
  apertura: boolean;
  aperturaAm: Date;
  aperturaPm: Date;
  cierreAm: Date;
  cierrePm: Date;
  horaAm: string;
  horaPm: string;
  editBarrio: null;
  editBrazoDominante: null;
  editCategoria: null;
  editEstatura: null;
  editFechaNacimiento: null;
  editGenero: null;
  editPeso: null;
  editTorneos: null;
  solBarrio: null;
  solBrazoDominante: null;
  solCategoria: null;
  solEstatura: null;
  solFechaNacimiento: null;
  solGenero: null;
  solPeso: null;
  solTorneos: null;
}